+++ app/adopt/seoul/[id]/_components/Bottom/AnimalVideo.tsx
