+++ app/adopt/seoul/[id]/_components/Top/AnimalPolaroid.tsx
@@ -2,8 +2,8 @@ import Image from 'next/image'
 import ImageSlide from './ImageSlide'
 import Cat from '/public/icons/cat.svg'
 import { useAtom } from 'jotai'
-import { animalAtom } from '../../_components/OneSeoulAnimal'
-import { mergedItem } from '../../_components/SeoulAnimalList'
+import { animalAtom } from '../../../_components/OneSeoulAnimal'
+import { mergedItem } from '../../../_components/SeoulAnimalList'
 import State from '@/constants/adoptState'
 import { colorMatch } from '@/constants/colorMatch'
