+++ app/adopt/seoul/[id]/page.tsx
@@ -1,9 +1,9 @@
 'use client'
 
 import { useRouter } from 'next/navigation'
-import AnimalDetail from './_components/AnimalDetail'
-import AnimalPolaroid from './_components/AnimalPolaroid'
-import AnimalInfoText from './_components/AnimalInfoText'
+import AnimalDetail from './_components/Bottom/AnimalDetail'
+import AnimalPolaroid from './_components/Top/AnimalPolaroid'
+import AnimalInfoText from './_components/Top/AnimalInfoText'
 
 const Detail: React.FC = () => {
   return (
